Fix login feature
User was having trouble logging in with either a username or email address. Additional security was added to login as well as validations to the user model. Username and email were indexed on the users table. The way a user is redirected to sign up was updated for user-friendliness. User are now redirected to the correct page once logged in.