Fix ability to send greetings to users
Problems with the previous commit are addressed. Users can now select a user whom he or she wants to send a message to, and the message will show up on his or her dashboard along with the time in which the greeting was sent. The most recent greetings will be shown first.